refactor(desktop): move the capture transport out of ConversationDetailView
ConversationDetailView is frozen by the product-file line-count ratchet
(1877 lines on main). The capture transport and its new scrubber are their
own surface, so they move to ChatFirst/ConversationCapturePlaybackSection.swift
next to the playback controller they drive. Pure move; no behavior change.
The detail view shrinks to 1765 lines. The new file joins the
capture-lifecycle e2e flow covers.

Verification:
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcrun swift test
  --package-path Desktop --filter 'CaptureArchiveTests|GlassContentChromeTests':
  37 tests, 0 failures
- swift-format lint on both files: clean
